Rename udpsocket file to socket file
