+++ src/lib.rs
@@ -29,10 +29,10 @@ extern crate gstreamer as gst;
 
 extern crate futures;
 extern crate tokio;
+extern crate tokio_current_thread;
 extern crate tokio_executor;
 extern crate tokio_reactor;
 extern crate tokio_threadpool;
-extern crate tokio_current_thread;
 extern crate tokio_timer;
 
 extern crate either;
@@ -46,7 +46,7 @@ extern crate net2;
 
 mod iocontext;
 
-mod udpsocket;
+mod socket;
 mod udpsrc;
 
 mod appsrc;
